+++ apps/api/src/manage/manage.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common";
+import { ManageController } from "./manage.controller.js";
+import { ManageService } from "./manage.service.js";
+
+@Module({
+  controllers: [ManageController],
+  providers: [ManageService],
+})
+export class ManageModule {}